Tweaks to add a regex filter.
I've also put in a note on the filename related options.  From what I can tell Plex is no longer passing media.filename to the agent, it keeps showing as None.  Since the agent is currently testing the filepath options against media.filename it doesn't seem to be working.

Also for some reason the search url to query TPDB wasn't correct, so I'm honestly not sure how it was working in the first place